+++ kafka/kstreams.go
@@ -62,7 +62,7 @@ func (k *Streams) Stream(ctx context.Context, handler ziggurat.Handler) error {
 func (k *Streams) stop() {
 	for _, consumers := range k.routeConsumerMap {
 		for i := range consumers {
-			k.Logger.Error("error stopping consumer: %v", consumers[i].Close())
+			k.Logger.Error("kstreams: error stopping consumer", consumers[i].Close())
 		}
 	}
 }